+++ workspace/ks-vm-new/src/types.rs
@@ -5,10 +5,10 @@ pub type Offset = i32;
 pub type VariableId = u32;
 pub type Pointer = usize;
 
-pub type Slot = u64;
+pub type Slot = u32;
 
-pub type CollectionId = u64;
-pub type StorageId = u64;
+pub type CollectionId = u32;
+pub type StorageId = u32;
 pub type Owners = u32;
 
 pub type NativeId = usize;
